// src/app/core/interfaces/carrousel.interface.ts
// Generated by https://quicktype.io

export interface CarrouselInfo {
    id:          number;
    title:       string;
    description: null | string;
    carrousel:   Carrousel[];
    button_text?:      string;
    button_url?:       string;
}

export interface Carrousel {
    id:          number;
    title:       string;
    description: string;
    image:       CarrouselImage;
}

export interface CarrouselImage {
    data: DAT;
}

export interface DAT {
    id:         number;
    attributes: FluffyAttributes;
}

export interface FluffyAttributes {
    name:              string;
    alternativeText:   null;
    caption:           null;
    width:             number;
    height:            number;
    formats:           Formats;
    hash:              string;
    ext:               EXT;
    mime:              MIME;
    size:              number;
    url:               string;
    previewUrl:        null;
    provider:          string;
    provider_metadata: null;
    createdAt:         string;
    updatedAt:         string;
}

export enum EXT {
    JPEG = ".jpeg",
    Jpg = ".jpg",
}

export interface Formats {
    thumbnail: Large;
    small?:    Large;
    medium?:   Large;
    large?:    Large;
}

export interface Large {
    name:   string;
    hash:   string;
    ext:    EXT;
    mime:   MIME;
    path:   null;
    width:  number;
    height: number;
    size:   number;
    url:    string;
}

export enum MIME {
    ImageJPEG = "image/jpeg",
}